Modifies player.points reset logic in infinity
Further testing showed that points were not resetting to below the
impossible-points threshold (i.e. 1e308) after bigCrunch().  To remedy
this, the code was changed to make the player.points reset occur as
soon as possible after the last time player.points is needed.
Additional calculations which would have pushed the points over the
impossible-points threshold _after_ a bigCrunch() now seem to avoid
triggering a repeated infinity.

It bears noting that this has only been tested with 3rd infinity, and
it's possible that ever-higher numbers would make even this fix
insufficient.  It would be prudent to test this fix with an endgame
save, to verify if the calculations hold up at that stage of the game.