Renames/refactors isCollision to doesOcclude
isCollision has been lifted higher and slightly optimized;
as notated in the new comments, the actual hit rate for exact-cover
occlusions is pretty low, such that it may be worth it to accept
the occasional total occlusion in lieu of paying the cost for
repeated occlusion checks.

That said, profiling while testing this code for optimizations also
revealed that the _true_ enemy here is the garbage collector;
investigation is forthcoming to see whether the GC is hanging on DOM
manipulation, event listener cleanup, both, neither, or other.