Hopefully squashing the challenge bug once and for all
Force all challenges to have the big crunch screen when reaching 1e308 points, and having the challenge clear when clicking the button. This makes it completely based on just reaching that screen, not based on any stats. This also removes a C5 bug where if it automatically resets between 1e300-1e308, it will kick you out without giving you the clear.