Wow, we were stupid huh
Apparently the issue was that the completeChallenge() function just presses the button on the challenge thing, it isn't a function to set the challenge to clear. This means that if your points are lower than 1e308 when that function is triggered, it will just kick you out of the challenge.